Updated the Final Project to ressolve the issue
I have changed the int values to float to increase the limit of salary also the statements are commented out on line 124, 125 and 163, 164 to resolve the issue of calculation of total tax.

+++ FinalProject--Python.py
@@ -195,9 +195,9 @@ def calculate():
 
     at = e5.get()
     ad = e6.get()
-    ta = int(at) - int(ad)
+    ta = float(at) - float(ad)
     old = oldtax(ta)
-    new = newtax(int(at))
+    new = newtax(float(at))
     tax_save = abs(new - old)
     tax_save = round(tax_save, 2)
     if new > old:
